chore(python): Make importable scripts non-executable.
As requested, scripts that can be imported, can not longer be used as
standalone executables.
